==> notebooks/1 Handling Missing Values.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import sklearn
import seaborn as sns

df = pd.read_csv("Train.csv")
df.head()

df.info()

df.nunique()

df.isnull().sum()

item_avg_weight = df.groupby(["Item_Identifier"])["Item_Weight"].mean()
item_avg_weight

miss_bool = df['Item_Weight'].isnull() 
miss_bool

miss_bool.sum()

df.loc[miss_bool,'Item_Identifier']

def impute_Item_Weight(df):
    item_avg_weight = df.groupby(["Item_Identifier"])["Item_Weight"].mean()
    item_avg_weight
    miss_bool = df['Item_Weight'].isnull() 
    print('Orignal #missing: %d'% sum(miss_bool))
    df.loc[miss_bool,'Item_Weight'] = df.loc[miss_bool,'Item_Identifier'].apply(lambda x: item_avg_weight.loc[x])
    print('Final #missing: %d'% sum(df['Item_Weight'].isnull()))

impute_Item_Weight(df)

miss_after_bool = df['Item_Weight'].isnull() 
miss_after_bool.sum()

miss_af = df.loc[miss_after_bool,'Item_Identifier']
miss_af

df.loc[miss_after_bool,:]

df.shape

df = df.loc[~miss_after_bool,:]

df.shape

import pandas as pd

def impute_Outlet_size(df):
    outlet_size_mode = df.pivot_table(values='Outlet_Size', columns='Outlet_Type', aggfunc=lambda x: x.mode()[0] if not x.mode().empty else None)
    print('Mode for each Outlet_Type:')
    print(outlet_size_mode)
    miss_bool = df['Outlet_Size'].isnull()
    for outlet_type in outlet_size_mode.columns:
        df.loc[df['Outlet_Type'] == outlet_type, 'Outlet_Size'] = df.loc[df['Outlet_Type'] == outlet_type, 'Outlet_Size'].fillna(outlet_size_mode[outlet_type])
    
    return df


impute_Outlet_size(df)

df.to_csv("Mid.csv",index = False)
