==> notebooks/3 Feature Engineering.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import sklearn
import seaborn as sns

df = pd.read_csv("Mid.csv")
df.head()

df.info()

df.isnull().sum()

df.nunique()

df.pivot_table(values='Item_Outlet_Sales',index='Outlet_Type')

def modify_item_visibility(df):
    visibility_avg = df.pivot_table(values='Item_Visibility', index='Item_Identifier')
    miss_bool = (df['Item_Visibility'] == 0)
    print('Number of 0 values initially: %d'%sum(miss_bool))
    df.loc[miss_bool,'Item_Visibility'] = df.loc[miss_bool,'Item_Identifier'].apply(lambda x: visibility_avg.loc[x])
    print('Number of 0 values after modification: %d'%sum(df['Item_Visibility'] == 0))

modify_item_visibility(df)

df["Item_Visibility"].describe()

df['Item_Identifier']

print('Frequency of Categories for varible Item Type')
df['Item_Type'].value_counts()

df.groupby(["Item_Identifier","Item_Type"]).size()

def broad_item_type(df):
    df['Item_Type_Combined'] = df['Item_Identifier'].apply(lambda x: x[0:2])
    df['Item_Type_Combined'] = df['Item_Type_Combined'].map({'FD':'Food',
                                                                 'NC':'Non-Consumable',
                                                                 'DR':'Drinks'})
    print(df['Item_Type_Combined'].value_counts())

broad_item_type(df)

df["Outlet_Establishment_Year"].value_counts()

df.groupby(["Outlet_Establishment_Year","Outlet_Identifier","Outlet_Type","Outlet_Location_Type"])["Item_Outlet_Sales"].mean()

def cal_outlet_year(df):
    df['Outlet_Years'] = 2013 - df['Outlet_Establishment_Year']
    print(df['Outlet_Years'].describe())

cal_outlet_year(df)

def modify_item_fat_content(data):
    print('Original Categories:')
    print(data['Item_Fat_Content'].value_counts())
    print('\nModified Categories:')
    data['Item_Fat_Content'] = data['Item_Fat_Content'].replace({'LF':'Low Fat',
                                                                 'reg':'Regular',
                                                                 'low fat':'Low Fat'})
    print(data['Item_Fat_Content'].value_counts())

modify_item_fat_content(df)

def non_consumable_category(data):
    data.loc[data['Item_Type_Combined']=="Non-Consumable",'Item_Fat_Content'] = "Non-Edible"
    print(data['Item_Fat_Content'].value_counts())

non_consumable_category(df)

def Item_Visibility_MeanRatio(data):
    visibility_item_avg = data.pivot_table(values='Item_Visibility',index='Item_Identifier')
    func = lambda x: x['Item_Visibility']/visibility_item_avg['Item_Visibility'][visibility_item_avg.index == x['Item_Identifier']][0]
    data['Item_Visibility'] = data.apply(func,axis=1).astype(float)
    data['Item_Visibility'].describe()
    

Item_Visibility_MeanRatio(df)

df.head()

df.shape

df.nunique()

from sklearn.preprocessing import LabelEncoder
def label_encoding(df):
    le = LabelEncoder()
    df['Outlet'] = le.fit_transform(df['Outlet_Identifier'])
    df['Outlet']
    var_mod = ['Item_Fat_Content','Outlet_Location_Type','Outlet_Size','Item_Type_Combined','Outlet_Type','Outlet']
    le = LabelEncoder()
    for i in var_mod:
        df[i] = le.fit_transform(df[i])

label_encoding(df)

df.head()

df.shape

def One_hot_encoding(df):
    df = pd.get_dummies(df, columns=['Item_Fat_Content','Outlet_Location_Type','Outlet_Size','Outlet_Type','Item_Type_Combined','Outlet'],drop_first = True)
    
    return df

df = One_hot_encoding(df)

df.head()

df.shape

df.dtypes

df.info()

df.to_csv("Final.csv",index = False)
